+++ datumaro/plugins/cvat_format/extractor.py
@@ -18,9 +18,8 @@ from .format import CvatPath
 
 class CvatExtractor(SourceExtractor):
     _SUPPORTED_SHAPES = ('box', 'polygon', 'polyline', 'points')
-    _NOTSET = object()
 
-    def __init__(self, path, subset=_NOTSET):
+    def __init__(self, path, subset=None):
         assert osp.isfile(path), path
         rootpath = osp.dirname(path)
         images_dir = ''
@@ -29,7 +28,7 @@ class CvatExtractor(SourceExtractor):
         self._images_dir = images_dir
         self._path = path
 
-        if subset is self._NOTSET:
+        if not subset:
             subset = osp.splitext(osp.basename(path))[0]
         super().__init__(subset=subset)
