+++ datumaro/util/__init__.py
@@ -4,8 +4,6 @@
 # SPDX-License-Identifier: MIT
 
 import attr
-import os
-import os.path as osp
 from contextlib import ExitStack
 from functools import partial, wraps
 from itertools import islice
@@ -13,32 +11,6 @@ from itertools import islice
 
 def find(iterable, pred=lambda x: True, default=None):
     return next((x for x in iterable if pred(x)), default)
-
-def dir_items(path, ext, truncate_ext=False):
-    items = []
-    for f in os.listdir(path):
-        ext_pos = f.rfind(ext)
-        if ext_pos != -1:
-            if truncate_ext:
-                f = f[:ext_pos]
-            items.append(f)
-    return items
-
-def split_path(path):
-    path = osp.normpath(path)
-    parts = []
-
-    while True:
-        path, part = osp.split(path)
-        if part:
-            parts.append(part)
-        else:
-            if path:
-                parts.append(path)
-            break
-    parts.reverse()
-
-    return parts
 
 def cast(value, type_conv, default=None):
     if value is None:
